pyproject.toml
[tool.poetry]
name = "car-fleet"
version = "0.1.0"
description = ""
authors = ["Denis <makrotheeast@gmail.com>"]
readme = "README.md"
packages = [{include = "car_fleet"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.1"
psycopg = "^3.1.9"
djangorestframework = "^3.14.0"
faker = "^18.10.1"
debugpy = "^1.6.7"
bandit = "^1.7.5"
autopep8 = "^2.0.2"


[tool.poetry.group.dev.dependencies]
yapf = "^0.33.0"
bandit = "^1.7.5"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
